fibonacci 4

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,16 +8,16 @@
 
 int main(void)
 {
-	int f1 = 0, f2 = 1, i, next;
+	long int f1 = 0, f2 = 1, i, next;
 
 	for (i = 1; i <= 98; i++)
 	{
 		next = f1 + f2;
 
 		if (i != 98)
-			printf("%d, ", next);
+			printf("%lu, ", next);
 		else
-			printf("%d\n", next);
+			printf("%lu\n", next);
 		f1 = f2;
 		f2 = next;
 	}
